requirements.txt
streamlit
pandas
numpy
plotly
langchain
langchain-openai
openai
python-dotenv
faiss-cpu
chromadb
sentence-transformers
langchain-community